improvement: move assets to image folder
