+++ packages/forma-36-website/src/components/Masthead.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { DisplayText, Paragraph } from '@contentful/forma-36-react-components';
 import tokens from '@contentful/forma-36-tokens';
 import { css } from '@emotion/core';
-import bg from './dot-bg.png';
+import bg from '../images/dot-bg.png';
 
 const styles = {
   masthead: css`
